Begin work on the command line.
